+++ json_serializable/lib/src/type_helpers/map_helper.dart
@@ -11,11 +11,6 @@ import '../utils.dart';
 
 const _keyParam = 'k';
 
-const _helperFunctionDefinition = '''T $_helperName<T extends Map>(T source) =>
-    (source == null || source.isEmpty) ? null : source;''';
-
-const _helperName = r'_$nullIfEmptyMap';
-
 class MapHelper extends TypeHelper<TypeHelperContextWithConfig> {
   const MapHelper();
 
@@ -36,18 +31,11 @@ class MapHelper extends TypeHelper<TypeHelperContextWithConfig> {
     final subFieldValue = context.serialize(valueType, closureArg);
     final subKeyValue = context.serialize(keyType, _keyParam);
 
-    final contextNullable = context.nullable || encodeEmptyAsNullRoot(context);
-
-    if (encodeEmptyAsNullRoot(context)) {
-      context.addMember(_helperFunctionDefinition);
-      expression = '$_helperName($expression)';
-    }
-
     if (closureArg == subFieldValue && _keyParam == subKeyValue) {
       return expression;
     }
 
-    final optionalQuestion = contextNullable ? '?' : '';
+    final optionalQuestion = context.nullable ? '?' : '';
 
     return '$expression$optionalQuestion'
         '.map(($_keyParam, $closureArg) => MapEntry($subKeyValue, $subFieldValue))';